requirements.txt
bleach==1.5.0
cycler==0.10.0
enum34==1.1.6
html5lib==0.9999999
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
numpy==1.14.5
protobuf==3.6.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
tensorflow-tensorboard==0.4.0
Werkzeug==0.14.1
